Add steps to compile protobuf definitions
This adds a build container definition and a docker compose file that
will allow you to easily compile your .proto into a .pb.go file,
without installing dependencies. The Dockerfile-build dockerfile
creates a container with the protobuf compiler and the compiler plugin
for go. The compose file sets this up to mount the current directory
and then compile the current directory leaving you with the compiled
sensorData.pb.go.